Can now route on unanswered checkboxes
FRONTEND - Added logic to be able to select 'unanswered' from the routing conditions list; doing so removes any previously chosen values for other routing conditions and removes the 'CHOOSE' button. Tests passing.

API - Added ability for the API to be able to recognise the 'unanswered' routing condition to save it in the DB. Tests passing.

PUBLISHER - Added logic to translate 'unanswered' into 'unset' so that Runner can understand what to do. Additionally, added logic to add appropriate routing to the JSON file and not show 'values' property if 'not set' is the routing condition. Tests passing.

Tested manually and it works.